% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{anaBiv}
\alias{anaBiv}
\title{anaBiv generic function}
\usage{
anaBiv(var, group, parallel, ...)
}
\arguments{
\item{var}{listVar object or data.frame}

\item{group}{Name of the factor variable to make subgroups with}

\item{parallel}{Logical. Make analysis using parallel from \code{\link[parallel:mclapply]{parallel::mclapply()}}.}

\item{...}{Further arguments: \code{normality} and \code{digits.p}, as in \code{\link[=descTab]{descTab()}}.}
}
\value{
A list of VarGroup object or data.frame
}
\description{
Generic function of anaBiv which gives bivariate analysis according to group
}
