% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_ana_biv.R
\name{ft_ana_biv}
\alias{ft_ana_biv}
\title{This function is depreciated, please use anaBiv(). \code{\link[=anaBiv]{anaBiv()}}}
\usage{
ft_ana_biv(...)
}
\arguments{
\item{...}{None}
}
\value{
No return value, depreciated
}
\description{
This function is depreciated, please use anaBiv(). \code{\link[=anaBiv]{anaBiv()}}
}
