% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_univ_tab.R
\name{ft_desc_tab}
\alias{ft_desc_tab}
\title{Generic function to create a table of descriptive analysis of a dataset}
\usage{
ft_desc_tab(
  data,
  group = NULL,
  complete = TRUE,
  quanti = FALSE,
  quali = FALSE,
  na.print = FALSE,
  p.value = TRUE,
  min.max = FALSE,
  digits.opt = 1,
  nonnormal = 0
)
}
\arguments{
\item{data}{A dataset, needs to be a dataframe object}

\item{group}{The variable of interest for you univariate analysis to sub-grouped by.}

\item{complete}{Wether to print quantitative and qualitative variables; default = TRUE}

\item{quanti}{To be used only if complete is FALSE. If TRUE, returns only the univariate analysis for quantitative variables}

\item{quali}{To be used only if complete is FALSE. If TRUE returns only the univariate analysis for qualitative variables}

\item{na.print}{Wether to print NAs n(\%), default = FALSE. Note that if true, "Total" will also be printed. This will be change in a futur version}

\item{p.value}{Print p value. Group needs to be set; default = TRUE. If TRUE, "Total" will also be printed}

\item{min.max}{Display min and max value for quantitative variables; default is false}

\item{digits.opt}{How many numbers after the "." you'd like for the proportions of qualitative variables; default is 1}

\item{nonnormal}{Treat all quantitative variable as non normal variables: will compute p.value according to Wilcox.test. Default is FALSE, TRUE will display med(IQR), 2 will display med(Q1 - Q3)}
}
\value{
The object returned depends on the "parse" option:either a dataframe or a kable oject
}
\description{
This function allows you to display all together all univariate analysis (min-max; median/mean; IQR/SD; proportions) and bivariates analysis (wilcoxon, chisq or fisher). The univariate analysis can be sub-grouped by a viariable of interest (binary variable)
}
\examples{
data(iris)
iris<-iris[iris$Species \%in\% c("virginica", "versicolor"),]
iris$Species<-droplevels(iris$Species)
ft_desc_tab(iris, group = "Species")
}
