% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_distrib.R
\name{ft_distrib}
\alias{ft_distrib}
\title{Graphical distribution of variables}
\usage{
ft_distrib(data)
}
\arguments{
\item{data}{An object of dataframe type}
}
\description{
Generic function to display graphical reprensentation of all data in a dataset
}
\examples{
library(doudpackage)
par(mfrow=c(3,3))
ft_distrib(cars)
}
