% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{initialize,parseClass-method}
\alias{initialize,parseClass-method}
\title{S4 class initialization function}
\usage{
\S4method{initialize}{parseClass}(
  .Object,
  table,
  group,
  pvalue,
  na.print,
  quanti,
  quali,
  var_list,
  data,
  digits.qt,
  digits.ql
)
}
\arguments{
\item{.Object}{The object to create}

\item{table}{The result of descTab}

\item{group}{The variable from which to make subgroups}

\item{pvalue, na.print, quanti, quali}{Values from descTab LINK}

\item{var_list}{An object of listVar LINK}

\item{data}{The dataset provided in descTab}

\item{digits.qt, digits.ql}{As provided in descTab}
}
\value{
parseClass object
}
\description{
Initialization function for parseClass object LINK
}
