% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{initialize,Var-method}
\alias{initialize,Var-method}
\title{S4 class initialization function}
\usage{
\S4method{initialize}{Var}(.Object, name, type, normal)
}
\arguments{
\item{.Object}{Object to be initialized}

\item{name}{A character taking name of the variable}

\item{type}{A character taking name of the variable type}

\item{normal}{Logical, if variable, is numeric; is it normal}
}
\value{
Var Object
}
\description{
Initialization function for Var \code{\link[=initialize,Var-method]{initialize,Var-method()}}
}
