% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{initialize,VarGroup-method}
\alias{initialize,VarGroup-method}
\title{S4 class initialization function}
\usage{
\S4method{initialize}{VarGroup}(
  .Object,
  x,
  group_var,
  pvalue,
  parsed_name,
  value,
  missing.value,
  missing.value.name
)
}
\arguments{
\item{.Object}{Object to be initialized}

\item{x}{A Var object}

\item{group_var}{The subgroup for whichproportions, mean/sd were calculated
and missing values}

\item{pvalue}{The calculated pvalue}

\item{parsed_name}{The name of the variable parsed with the n (\%), mean (SD)}

\item{value}{The values calculated parsed}

\item{missing.value}{Missing values numbers and proportions n (\%)}

\item{missing.value.name}{Missing values concatenate with the leve
of the variable if it factor}
}
\value{
VarGroup object
}
\description{
Initialization function for VarGroup \code{\link[=initialize,VarGroup-method]{initialize,VarGroup-method()}}
}
