% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{parseClass-class}
\alias{parseClass-class}
\title{S4 class}
\description{
A S4 class containing all the information needed for parsClassFun
the missing values and the group for which it was calculated
}
\section{Slots}{

\describe{
\item{\code{table}}{The result of descTab}

\item{\code{group}}{The variable from which to make subgroups}

\item{\code{pvalue,na.print,quanti,quali}}{Values from descTab \code{\link[=descTab]{descTab()}}}

\item{\code{var_list}}{An object of listVar \code{\link[=listVar-class]{listVar-class()}}}

\item{\code{data}}{The dataset provided in descTab}

\item{\code{digits.qt,digits.ql}}{As provided in descTab}
}}

