% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{parseClass}
\alias{parseClass}
\title{S4 class initialization function}
\usage{
parseClass(
  table,
  group,
  pvalue,
  na.print,
  quanti,
  quali,
  var_list,
  data,
  digits.qt,
  digits.ql
)
}
\arguments{
\item{table}{The result of descTab}

\item{group}{The variable from which to make subgroups}

\item{pvalue, na.print, quanti, quali}{Values from descTab \code{\link[=descTab]{descTab()}}}

\item{var_list}{An object of listVar \code{\link[=listVar-class]{listVar-class()}}}

\item{data}{The dataset provided in descTab}

\item{digits.qt, digits.ql}{As provided in descTab}
}
\value{
parseClass object
}
\description{
Initialization function for parseClass object \code{\link[=initialize,parseClass-method]{initialize,parseClass-method()}}
}
