% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseClassFun,parseClass-method}
\alias{parseClassFun,parseClass-method}
\title{Make the LaTeX/HTML table}
\usage{
\S4method{parseClassFun}{parseClass}(
  table,
  col.order = NULL,
  levels_to_keep = NULL,
  group_rows_labels = NULL,
  font = "arial"
)
}
\arguments{
\item{table}{The output of \code{\link[=descTab]{descTab()}} or \code{\link[=anaBiv]{anaBiv()}}, an S4 object.}

\item{col.order}{Optional. A vector containing the column order. If set, must contains at least all levels of group. Three columns created are "var", "Total", and "pvalue" which can be present in the vector}

\item{levels_to_keep}{Optional, named list. If the variable is binary, which level to keep. Default is the last level of levels(variable). Must be as: list("variable name" = "level to keep").}

\item{group_rows_labels}{Optional, named list. Create row labels in order to regroup them. Must be as list("label" = c("var1", "var2), "label2" = c("var3", "var4")). Rows are reordered to follow the order of the labels and of the variables within each label.}

\item{font}{The HTML font of the table. Default "arial".}
}
\value{
An HTML/LaTex file which can be used directly in Rmarkdown and copy paste
}
\description{
This functions takes the S4 output of descTab to create an HTML parsed table
}
\examples{
data(iris)
iris$fact_1 <- as.factor(sample(1:5, 150, replace = TRUE))
iris$fact_1[sample(1:150, 30)] <- NA
iris$fact_2 <- as.factor(sample(c("A", "B"), 150, replace = TRUE))
iris$num <- runif(150, min = 0, max = 100)
iris$num[sample(1:150, 5)] <- NA
iris_test <- descTab(iris, group = "Species", na.print = TRUE)

# fact_2 is binary: keep the "A" line only, and group rows under two labels
parseClassFun(iris_test,
              levels_to_keep = list("fact_2" = "A"),
              group_rows_labels = list("Size" = c("Petal.Length", "Petal.Width"),
                                       "My_f" = c("num", "fact_2")))
}
\seealso{
\code{\link[=descTab]{descTab()}} which produces the object this function renders.
}
