% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{[,Var-method}
\alias{[,Var-method}
\title{Method to access S4 Var elements}
\usage{
\S4method{[}{Var}(x, i)
}
\arguments{
\item{x}{: object}

\item{i}{: value}
}
\value{
object of Var
}
\description{
Method to access Var elements by name
}
