% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{[,VarGroup-method}
\alias{[,VarGroup-method}
\title{Method to access S4 Var elements}
\usage{
\S4method{[}{VarGroup}(x, i)
}
\arguments{
\item{x}{: object}

\item{i}{: value}
}
\value{
object element
}
\description{
Mathod to access VarGroup \code{\link[=initialize,VarGroup-method]{initialize,VarGroup-method()}} elements by name
}
