% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{[<-,parseClass,character,ANY,ANY-method}
\alias{[<-,parseClass,character,ANY,ANY-method}
\title{Method to modify S4 Var elements}
\usage{
\S4method{[}{parseClass,character,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{: Object}

\item{i}{: Element name}

\item{value}{: Value to be added}
}
\value{
object
}
\description{
Method to modify parseClass \code{\link[=initialize,parseClass-method]{initialize,parseClass-method()}} elements by name
}
