% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{[<-,parseClass-method}
\alias{[<-,parseClass-method}
\title{Method to access S4 Var elements}
\usage{
\S4method{[}{parseClass}(x, i) <- value
}
\arguments{
\item{x}{: Object}

\item{i}{: Element name}

\item{value}{: Value to be added}
}
\value{
object
}
\description{
Method to modify VarGroup LINK elements by name
}
