% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{[<-,Var-method}
\alias{[<-,Var-method}
\title{Method to access S4 Var elements}
\usage{
\S4method{[}{Var}(x, i) <- value
}
\arguments{
\item{x}{: object}

\item{i}{: Element name}

\item{value}{: Value to be added}
}
\value{
object
}
\description{
Method to modify Var elements by name
}
