% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Var-class}
\alias{Var-class}
\title{S4 class}
\description{
A S4 class containing name, type and normality assessment of variable
}
\section{Slots}{

\describe{
\item{\code{name}}{A character taking name of the variable}

\item{\code{type}}{A character taking name of the variable type}

\item{\code{normal}}{Logical, if variable, is numeric; is it normal}
}}

