% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{Var}
\alias{Var}
\title{S4 class initialization function}
\usage{
Var(name, type = "", normal = TRUE)
}
\arguments{
\item{name}{A character taking name of the variable}

\item{type}{A character taking name of the variable type}

\item{normal}{Logical, if variable, is numeric; is it normal}
}
\value{
Var Object
}
\description{
Initialization function for Var LINK
}
