% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{VarGroup-class}
\alias{VarGroup-class}
\title{S4 class}
\description{
A S4 class containing Var \code{\link[=initialize,Var-method]{initialize,Var-method()}} It also contains the pvalue, the parsed value
the missing values and the group for which it was calculated
}
\section{Slots}{

\describe{
\item{\code{group_var}}{The subgroup for which proportions, mean/sd were calculated
and missing values}

\item{\code{pvalue}}{The calculated pvalue}

\item{\code{parsed_name}}{The name of the variable parsed with the n (\%), mean (SD)}

\item{\code{value}}{The values calculated parsed}

\item{\code{missing.value}}{Missing values numbers and proportions n (\%)}

\item{\code{missing.value.name}}{Missing values concatenate with the level
of the variable if it factor}
}}

