% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{anaBiv}
\alias{anaBiv}
\title{anaBiv generic function}
\usage{
anaBiv(var, group, ...)
}
\arguments{
\item{var}{listVar object or data.frame}

\item{group}{Variable to make subgroups with}

\item{...}{digits.p can be specified as descTab}
}
\value{
A list of VarGroup object or data.frame
}
\description{
Generic function of anaBiv which gives bivariate analysis according to group
}
