% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_ana_biv.R
\name{ft_ana_biv}
\alias{ft_ana_biv}
\title{Bivariate analysis}
\usage{
ft_ana_biv(data, group, signi = 0.2)
}
\arguments{
\item{data}{A database, dataframe object}

\item{group}{Variable from which sub-groups are made to compute p values}

\item{signi}{Significance levels (def = 0.20)}
}
\value{
Return a dataframe with univariate analaysis
}
\description{
Function that allows you to compute a p value according to a binary variable; default tests are Wilcoxon for quantitative variables, chisq or fisher wether chisq.test returns a warning
}
