% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseClassFun,parseClass-method}
\alias{parseClassFun,parseClass-method}
\title{Make the LaTeX/HTML table}
\usage{
\S4method{parseClassFun}{parseClass}(
  table,
  col.order = NULL,
  levels_to_keep = NULL,
  group_rows_labels = NULL
)
}
\arguments{
\item{table}{The output of \code{\link[=descTab]{descTab()}} or \code{\link[=anaBiv]{anaBiv()}}, an S4 object.}

\item{col.order}{Optional. A vector containing the column order. If set, must contains at least all levels of group. Three columns created are "var", "Total", and "pvalue" which can be present in the vector}

\item{levels_to_keep}{Optional, named list. If the variable is binary, which level to keep. Default is the last level of levels(variable). Must be as: list("variable name" = "level to keep").}

\item{group_rows_labels}{Optional, named list. Create row labels in order to regroup them. Must be as list("label" = c("var1", "var2), "label2" = c("var3", "var4")).}
}
\description{
This functions takes the S4 output of descTab to create an HTML parsed table
}
